tests/e2e/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
pytest==7.4.3
pytest-asyncio==0.23.2 